use float64 for listing coordinates

+++ packages/service-listing/internal/adapter/controller/apigateway.go
@@ -51,7 +51,7 @@ func (h *ApiGatewayHandler) FindListingsWithinLatLng(event events.APIGatewayProx
 	if err != nil {
 		return FindListingsWithinLatLngResponsePayload{}, customerror.ErrorHandler.ConvertError("distance", "String", err)
 	}
-	listings, err := h.service.FindListingsWithinLatLng(float32(latitude), float32(longitude), int32(distance))
+	listings, err := h.service.FindListingsWithinLatLng(latitude, longitude, int32(distance))
 	return FindListingsWithinLatLngResponsePayload{
 		Listings: listings,
 	}, err.(*customerror.CustomError)
@@ -61,8 +61,8 @@ func (h *ApiGatewayHandler) AddListing(event events.APIGatewayProxyRequest) (Add
 	body := struct {
 		LenderId      string   `json:"lenderId"`
 		StreetAddress string   `json:"streetAddress"`
-		Latitude      float32  `json:"latitude"`
-		Longitude     float32  `json:"longitude"`
+		Latitude      float64  `json:"latitude"`
+		Longitude     float64  `json:"longitude"`
 		ImageUrls     []string `json:"imageUrls"`
 		Title         string   `json:"title"`
 		FeeAmount     int32    `json:"feeAmount"`
